apns: filters: move FilteredStream
